+++ src/libopenrave/kinbodylink.cpp
@@ -147,6 +147,61 @@ void KinBody::Link::SetMass(dReal mass)
     GetParent()->_ParametersChanged(Prop_LinkDynamics);
 }
 
+AABB KinBody::Link::ComputeLocalAABB() const
+{
+    if( _vGeometries.size() == 1) {
+        return _vGeometries.front()->ComputeAABB(Transform());
+    }
+    else if( _vGeometries.size() > 1 ) {
+        Vector vmin, vmax;
+        bool binitialized=false;
+        AABB ab;
+        FOREACHC(itgeom,_vGeometries) {
+            ab = (*itgeom)->ComputeAABB(Transform());
+            if( ab.extents.x <= 0 || ab.extents.y <= 0 || ab.extents.z <= 0 ) {
+                continue;
+            }
+            Vector vnmin = ab.pos - ab.extents;
+            Vector vnmax = ab.pos + ab.extents;
+            if( !binitialized ) {
+                vmin = vnmin;
+                vmax = vnmax;
+                binitialized = true;
+            }
+            else {
+                if( vmin.x > vnmin.x ) {
+                    vmin.x = vnmin.x;
+                }
+                if( vmin.y > vnmin.y ) {
+                    vmin.y = vnmin.y;
+                }
+                if( vmin.z > vnmin.z ) {
+                    vmin.z = vnmin.z;
+                }
+                if( vmax.x < vnmax.x ) {
+                    vmax.x = vnmax.x;
+                }
+                if( vmax.y < vnmax.y ) {
+                    vmax.y = vnmax.y;
+                }
+                if( vmax.z < vnmax.z ) {
+                    vmax.z = vnmax.z;
+                }
+            }
+        }
+        if( !binitialized ) {
+            ab.pos = _t.trans;
+            ab.extents = Vector(0,0,0);
+        }
+        else {
+            ab.pos = (dReal)0.5 * (vmin + vmax);
+            ab.extents = vmax - ab.pos;
+        }
+        return ab;
+    }
+    return AABB();
+}
+
 AABB KinBody::Link::ComputeAABB() const
 {
     if( _vGeometries.size() == 1) {
@@ -158,7 +213,7 @@ AABB KinBody::Link::ComputeAABB() const
         AABB ab;
         FOREACHC(itgeom,_vGeometries) {
             ab = (*itgeom)->ComputeAABB(_t);
-            if((ab.extents.x == 0)&&(ab.extents.y == 0)&&(ab.extents.z == 0)) {
+            if( ab.extents.x <= 0 || ab.extents.y <= 0 || ab.extents.z <= 0 ) {
                 continue;
             }
             Vector vnmin = ab.pos - ab.extents;